✨ Fix dashboard
