+++ src/app/common/getUser.ts
@@ -1,0 +1,8 @@
+//supabaseからユーザー情報を取得したい
+import { createClient } from '@/app/utils/supabase/server';
+
+export const getUser = async () => {
+  const supabase = await createClient();
+  const { data: user } = await supabase.auth.getUser();
+  return user;
+};
