+++ src/app/dashboard/components/Card.tsx
@@ -27,10 +27,26 @@ const Card = ({ event, color }: CardProps) => {
           }`}
         >
           <h3 className='font-semibold text-lg truncate'>{event.name}</h3>
-          {/* <p className='text-sm text-gray-500'>
-          {event.dateOptions[0].toLocaleString()}
-        </p>
-        <p className='text-sm text-gray-400'>{event.respondents} respondents</p> */}
+          <p className='text-sm text-gray-500'>
+            {new Date(
+              // タイムスロットを日付でソートして最新のものを取得
+              [...event.time_slots].sort(
+                (a, b) =>
+                  new Date(b.event_date).getTime() -
+                  new Date(a.event_date).getTime()
+              )[0].event_date
+            ).toLocaleDateString('en-US', {
+              weekday: 'short',
+              month: 'short',
+              day: 'numeric',
+            })}
+            {' | '} {event.time_slots[0].time}
+            {
+              //event.time_slotsが2つ以上の場合は and more 残りの時間スロットを表示
+              event.time_slots.length > 1 &&
+                ` and ${event.time_slots.length - 1} more`
+            }
+          </p>
         </div>
       </Transition>
     </div>
